requirements.txt
celery>=5.5.3
chromadb>=0.5.4
fastapi>=0.120.0
jwt>=1.4.0
langchain>=1.0.2
openai>=2.6.0
passlib>=1.7.4
pdfplumber>=0.11.7
pydantic>=2.12.3
pytesseract>=0.3.13
python-docx>=1.2.0
python-multipart>=0.0.20
redis>=7.0.0
sqlalchemy>=2.0.44
uvicorn[standard]>=0.38.0
